Harden locator test coverage and split the docs contract tests
Address the third review round:

- Assert every candidate the locator generates in the missing-binary
  test, including the `CARGO_TARGET_DIR/<triple>/<profile>/` path.
- Add a both-present priority test so inverting the candidate order
  can no longer pass unnoticed.
- Cover `test_support::fs::try_is_file` directly: a directory reports
  `Ok(false)`, and traversal through a regular file propagates the
  error rather than being folded into absence.
- Pass the located `Utf8PathBuf` straight to
  `assert_cmd::Command::new`, completing the camino migration.
- Split the installation and release-contract tests (149 lines with
  no entanglement) into `tests/documentation_installation_tests.rs`,
  bringing `documentation_examples_tests.rs` from 505 lines to 355,
  under the repository's 400-line file cap. The new binary reuses the
  shared `documentation_examples` directory module via the same
  `pub mod` pattern as the loader tests; behaviour is unchanged.
- Document `try_is_file` in the developers' guide under
  `test_support::fs`, with a cross-reference from the binary-locator
  section.
